hemorrhage/requirements.txt
PySocks==1.7.1
brotlipy==0.7.0
cffi==1.14.6
chardet==4.0.0
conda-package-handling==1.7.3
cryptography==3.4.7
idna==2.10
pip==21.1.3
pyOpenSSL==20.0.1
pycosat==0.6.3
pycparser==2.20
requests==2.25.1
ruamel-yaml-conda==0.15.100
setuptools==52.0.0.post20210125
six==1.16.0
tqdm==4.61.2
wheel==0.36.2
GitPython==3.1.43
MarkupSafe==2.1.3
PyYAML==6.0.1
boltons==23.0.0
certifi==2024.7.4
click==8.1.7
conda==23.5.2
docker-pycreds==0.4.0
filelock==3.13.1
gitdb==4.0.11
gmpy2==2.1.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.22.3
opencv-python==4.7.0.72
packaging==24.1
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.10.1
sentry-sdk==2.14.0
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
triton==2.1.0
typing_extensions==4.11.0
urllib3==1.26.20
wandb==0.18.1
Babel==2.10.3
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.2.0
Pygments==2.12.0
Send2Trash==1.8.0
Werkzeug==2.2.2
absl-py==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
fonttools==4.34.4
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-language-pack-zh-CN==3.4.post3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
widgetsnbextension==3.6.1
zipp==3.8.1
